+++ src/resolvers/versionedSection.ts
@@ -1,6 +1,6 @@
-import { PublishedSectionSearchResults, Resolvers } from "../types";
+import { Resolvers, VersionedSectionSearchResults } from "../types";
 import { MyContext } from "../context";
-import { VersionedSection } from "../models/VersionedSection";
+import { VersionedSection, VersionedSectionSearchResult } from "../models/VersionedSection";
 import { Section } from "../models/Section";
 import { Tag } from "../models/Tag";
 import { VersionedTemplate } from "../models/VersionedTemplate";
@@ -35,21 +35,22 @@ export const resolvers: Resolvers = {
       }
     },
     // Get all of the published versionedSections with the given name
-    publishedSections: async (_, { term, paginationOptions }, context: MyContext): Promise<PublishedSectionSearchResults> => {
+    publishedSections: async (_, { term, paginationOptions }, context: MyContext): Promise<VersionedSectionSearchResults> => {
       const reference = 'publishedSections resolver';
       try {
         const opts = !isNullOrUndefined(paginationOptions) && paginationOptions.type === PaginationType.OFFSET
                     ? paginationOptions as PaginationOptionsForOffsets
                     : { ...paginationOptions, type: PaginationType.CURSOR } as PaginationOptionsForCursors;
 
-        return await VersionedSection.findByName(reference, context, term, opts);
+        // Find published versionedSections with similar names for the current user
+        return await VersionedSectionSearchResult.search(reference, context, term, opts);
       } catch (err) {
         if (err instanceof GraphQLError) throw err;
 
         formatLogMessage(context).error(err, `Failure in ${reference}`);
         throw InternalServerError();
       }
-    }
+    },
   },
 
   VersionedSection: {