+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Added
 - Added `queryWithPagination` function to `MySQLModel`.
+- Added `VersionedSectionSearchResult` to optimize the query for displaying an org's published sections and the bestPractice sections
 - Added `popularFunders` query to the `affiliations` resolver
 - Added `accessLevel` to projectCollaborator and removed `userId`
 - Added new resolvers related to `projectCollaborators`. Also, when project is created, automatically add user as `projectCollaborator` with `access level`= `OWN`
@@ -78,6 +79,7 @@
 
 ### Updated
 - Updated `publishedTemplates`, `users`, `myTemplates`, `topLevelResearchDomains`, `repositories`, `myProjects`, `metadataStandards`, `licenses`, `affiliations` queries to use the new `paginationService`
+- Updated `publishedSections` resolver to return the new `VersionedSectionSearchResult` array
 - Format ORCID identifiers consistently, in the `Contributor` and `User` models, the `projectImport` resolver and `orcid` scalar.
 - Changed a number of GraphQL definitions to PascalCase.
 - Fixed projectCollaborators table which had an FKey on the plans table instead of the projects table
@@ -126,6 +128,7 @@
 - added bestPractice flag to the Section
 
 ### Removed
+- Removed duplicate section check from `Section.create` that was just going off of the "name"
 - Dropped the `PlanVersion` table
 - Removed `prepareAPITarget` function.
 - Removed `id` from `Project` model's constructor (already handled in base `MySQLModel`)
@@ -134,6 +137,7 @@
 
 ### Fixed
 - Fixed an issue where adding `templateCollaborators` was failing due to the fact that the `userId` field was required.
+-Was getting `undefined` bestPractice in `updateTemplate` when none was passed in because of the logic on how it was set. Added a check for whether `bestPractice` is defined before setting value. Also, added an update to `createTemplateVersion`, so that errors from the `generateTemplateVersion` will be caught and passed back in graphql response. Previously, when trying to save a DRAFT of a template, the mutation wouldn't return an error to the client, even though the `Save draft` did not successfully complete. [#265]
 - Removed `Copy of` from in front of copied `Section` and `Template` names [#261]
 - Fixed an issue where adding `templateCollaborators` was failing due to the fact that the `userId` field was required.
 - Fixed an issue where adding `projectCollaborators` was failing due to the fact that the `userId` field was required. This should not be required to add a new collaborator [#260]